requirements/common.txt
Django==1.10.1
django-custom-user==0.6
dj-database-url==0.4.1
psycopg2==2.6.1
django-rest-knox==2.2.0
django-cors-headers==1.1.0
django-filter==0.15.3
boto==2.42.0
Pillow==3.3.1
pycrypto==2.6.1
django-countries==4.0
django-localflavor==1.3
drf-extensions==0.3.1
django-phonenumber-field==1.1.0
